+++ src/main/java/alfio/manager/ExtensionManager.java
@@ -21,6 +21,7 @@ import alfio.extension.ExtensionService;
 import alfio.model.*;
 import alfio.model.extension.InvoiceGeneration;
 import alfio.repository.EventRepository;
+import alfio.repository.TicketReservationRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -32,6 +33,7 @@ public class ExtensionManager {
 
     private final ExtensionService extensionService;
     private final EventRepository eventRepository;
+    private final TicketReservationRepository ticketReservationRepository;
 
     public enum ExtensionEvent {
         RESERVATION_CONFIRMED,
@@ -129,6 +131,7 @@ public class ExtensionManager {
 
         Map<String, Object> payload = new HashMap<>();
         payload.put("reservationIds", reservationIds);
+        payload.put("reservations", ticketReservationRepository.findByIds(reservationIds));
 
         syncCall(extensionEvent, event, organizationId, payload, Boolean.class);
     }