+++ src/main/java/alfio/util/WorkingDaysAdjusters.java
@@ -20,7 +20,9 @@ import lombok.experimental.UtilityClass;
 
 import java.time.DayOfWeek;
 import java.time.LocalTime;
-import java.time.temporal.*;
+import java.time.temporal.ChronoUnit;
+import java.time.temporal.Temporal;
+import java.time.temporal.TemporalAdjuster;
 import java.util.*;
 
 @UtilityClass
@@ -31,6 +33,10 @@ public class WorkingDaysAdjusters {
 
     public static TemporalAdjuster defaultWorkingDays() {
         return temporal -> adjust(temporal, MON_FRI, ALL_DAY);
+    }
+
+    public static TemporalAdjuster workingDaysAtNoon() {
+        return temporal -> adjust(temporal, MON_FRI, Collections.singletonList(new HoursRange(LocalTime.of(12, 0, 0), LocalTime.of(12, 0, 0))));
     }
 
     private static Temporal adjust(Temporal in, Set<DayOfWeek> dayOfWeeks, List<HoursRange> hoursRanges) {