+++ src/main/webapp/resources/js/admin/feature/extension/extension.service.js
@@ -6,26 +6,23 @@
 
     function ExtensionService($http, HttpErrorHandler, EventService, OrganizationService, $q) {
         return {
-            loadEventConfig: function(organizationName, eventShortName) {
-                return $http.get('/admin/api/extensions/setting/organization/'+encodeURIComponent(organizationName)+'/event/'+encodeURIComponent(eventShortName)).error(HttpErrorHandler.handle);
+            loadEventConfig: function(orgId, eventShortName) {
+                return $http.get('/admin/api/extensions/setting/organization/'+orgId+'/event/'+encodeURIComponent(eventShortName)).error(HttpErrorHandler.handle);
             },
-            loadOrganizationConfig: function(organizationName) {
-                return $http.get('/admin/api/extensions/setting/organization/'+encodeURIComponent(organizationName)).error(HttpErrorHandler.handle);
+            loadOrganizationConfig: function(orgId) {
+                return $http.get('/admin/api/extensions/setting/organization/'+orgId).error(HttpErrorHandler.handle);
             },
             loadSystem: function() {
                 return $http.get('/admin/api/extensions/setting/system').error(HttpErrorHandler.handle);
             },
             loadEventConfigWithOrgIdAndEventId: function(orgId, eventId) {
                 var service = this;
-                return $q.all([OrganizationService.getOrganization(orgId), EventService.getEventById(eventId)]).then(function(res) {
-                    return service.loadEventConfig(res[0].data.name, res[1].data.shortName);
+                return $q.all([EventService.getEventById(eventId)]).then(function(res) {
+                    return service.loadEventConfig(orgId, res[0].data.shortName);
                 })
             },
             loadOrganizationConfigWithOrgId: function (orgId) {
-                var service = this;
-                return OrganizationService.getOrganization(orgId).then(function(org) {
-                    return service.loadOrganizationConfig(org.data.name);
-                });
+                return this.loadOrganizationConfig(orgId);
             },
 
             saveBulkSystemSetting: function(toSave) {
@@ -33,33 +30,25 @@
             },
 
             saveBulkOrganizationSetting: function(orgId, toSave) {
-                return OrganizationService.getOrganization(orgId).then(function(org) {
-                    var organizationName = org.data.name;
-                    return $http.post('/admin/api/extensions/setting/organization/' + encodeURIComponent(organizationName) + '/bulk-update', transformSettingPayload(toSave));
-                });
+                return $http.post('/admin/api/extensions/setting/organization/' + orgId + '/bulk-update', transformSettingPayload(toSave));
             },
 
             saveBulkEventSetting: function(orgId, eventId, toSave) {
-                return $q.all([OrganizationService.getOrganization(orgId), EventService.getEventById(eventId)]).then(function(res) {
-                    var organizationName = res[0].data.name;
-                    var eventShortName = res[1].data.shortName;
-                    return $http.post('/admin/api/extensions/setting/organization/'+encodeURIComponent(organizationName)+'/event/'+encodeURIComponent(eventShortName) + '/bulk-update', transformSettingPayload(toSave));
+                return $q.all([EventService.getEventById(eventId)]).then(function(res) {
+                    var eventShortName = res[0].data.shortName;
+                    return $http.post('/admin/api/extensions/setting/organization/'+orgId +'/event/'+encodeURIComponent(eventShortName) + '/bulk-update', transformSettingPayload(toSave));
                 });
             },
             deleteSystemSettingValue: function(toDelete) {
                 return $http.delete('/admin/api/extensions/setting/system/'+toDelete.id);
             },
             deleteOrganizationSettingValue: function(orgId, toDelete) {
-                return OrganizationService.getOrganization(orgId).then(function(org) {
-                    var organizationName = org.data.name;
-                    return $http.delete('/admin/api/extensions/setting/organization/' + encodeURIComponent(organizationName) + '/'+toDelete.id);
-                });
+                return $http.delete('/admin/api/extensions/setting/organization/' + orgId + '/'+toDelete.id);
             },
             deleteEventSettingValue: function(orgId, eventId, toDelete) {
-                return $q.all([OrganizationService.getOrganization(orgId), EventService.getEventById(eventId)]).then(function(res) {
-                    var organizationName = res[0].data.name;
-                    var eventShortName = res[1].data.shortName;
-                    return $http.delete('/admin/api/extensions/setting/organization/'+encodeURIComponent(organizationName)+'/event/'+encodeURIComponent(eventShortName) + '/'+toDelete.id);
+                return $q.all([EventService.getEventById(eventId)]).then(function(res) {
+                    var eventShortName = res[0].data.shortName;
+                    return $http.delete('/admin/api/extensions/setting/organization/'+orgId+'/event/'+encodeURIComponent(eventShortName) + '/'+toDelete.id);
                 });
             }
         };